+++ tests/utils.py
@@ -107,7 +107,7 @@ SIM_FULL = Simulation(
             medium=PoleResidue(eps_inf=1.0, poles=((6206417594288582j, (-3.311074436985222e16j)),)),
         ),
         Structure(
-            geometry=CustomSurfaceMeshGeometry.from_triangles(
+            geometry=TriangleMesh.from_triangles(
                 np.array(
                     [
                         [[1, 0, 0], [0, 1, 0], [0, 0, 1]],