+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `TriangleMesh` class for modeling geometries specified by triangle surface meshes, with support for STL file import.
 
 ### Changed